requirements.txt
oauth2client==4.1.3
gspread==6.1.4
python-telegram-bot==21.10